requirements.txt
asgiref==3.7.2
boto3==1.34.4
botocore==1.34.4
click==8.1.7
dj-database-url==2.1.0
Django==4.2.5
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-mptt==0.15.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
h11==0.14.0
inflection==0.5.1
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.25.0
whitenoise==6.6.0
